+++ src/App.tsx
@@ -1,9 +1,12 @@
 import { useEffect } from 'react'
+import { httpOnlyProxy } from './env'
 
 import { useLoadCredentialsFromCookies } from './hooks/authHooks'
 import { AppRoutes } from './routes'
 
 function App() {
+  console.log('httpOnlyProxy -->', httpOnlyProxy)
+  
   const { loadCredentials } = useLoadCredentialsFromCookies()
 
   useEffect(() => {
